[MGF] Enable reading MS level

+++ 3rdparty/libmgf/src/FlexLexer.h
@@ -149,8 +149,8 @@ extern "C++"
         virtual int yywrap();
 
     protected:
-        virtual int LexerInput(char* buf, int max_size);
-        virtual void LexerOutput(const char* buf, int size);
+        virtual size_t LexerInput(char* buf, size_t max_size);
+        virtual void LexerOutput(const char* buf, size_t size);
         virtual void LexerError(const char* msg);
 
         void yyunput(int c, char* buf_ptr);
